+++ backend/app/helpers/__init__.py
@@ -1,6 +1,6 @@
 from .sender import send_message
 from .vcard import extract_phone_number
-from .audio import transcribe_audio, summarize_transcript, transcript_to_base64, get_journal_goal_id, process_journal
+from .audio import transcribe_audio, process_journal
 from .text import (
     extract_emoji,
     is_valid_rating_digits,
@@ -16,15 +16,12 @@ __all__ = [
     "send_message",
     "extract_phone_number",
     "transcribe_audio",
-    "summarize_transcript",
     "extract_emoji",
     "is_valid_rating_digits",
     "get_monday_before",
     "look_back_summary",
     "parse_time_string",
     "get_timezone_from_number",
-    "transcript_to_base64",
-    "get_journal_goal_id",
     "get_timezone_safe",
     "get_goals_not_tracked_today",
     "process_journal",
