+++ backend/app/helpers/audio/__init__.py
@@ -1,3 +1,4 @@
-from .journaling import transcribe_audio, summarize_transcript, transcript_to_base64, get_journal_goal_id, process_journal
+from .transcribe_audio import transcribe_audio
+from .journaling import process_journal
 
-__all__ = ["transcribe_audio", "summarize_transcript", "transcript_to_base64", "get_journal_goal_id", "process_journal"]
+__all__ = ["transcribe_audio", "process_journal"]
