+++ backend/app/helpers/webhook/response_builder.py
@@ -11,10 +11,10 @@ def error_response(message: str, data=None, status_code: int = 400) -> tuple[Res
     }), status_code
 
 
-def success_response(data: dict, message: str = None, status_code: int = 200) -> tuple[Response, int]:
+def success_response(message: str = None, status_code: int = 200, **kwargs) -> tuple[Response, int]:
     """Build a success response body."""
     return jsonify({
         "success": True,
         "message": message,
-        "data": data,
+        "data": kwargs,
     }), status_code
